Fix data type for array length and update sum calculation

+++ Difficulty: Easy/Missing in Array/missing-in-array.java
@@ -1,7 +1,7 @@
 class Solution {
     int missingNum(int arr[]) {
         // code here
-        long  n=arr.length;
+        int  n=arr.length;
         long sum=0;
         
         //sum of elements of array
@@ -9,7 +9,7 @@ class Solution {
             sum=sum+arr[i];
         }
         
-        long Tsum=((n+1)*(n+2))/2; //sum of n natural number
+        long Tsum= (((long)(n+1)*(n+2))/2); //sum of n natural number
         
         return (int)(Tsum-sum);
     }
